Array iterat

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -7,9 +7,9 @@
  *@size: Size of array
  *@action: Pointer to function to execute
  */
-void array_iterator(int *array, size_t size, void(*action)(int))
+void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 
 	for (i = 0; i < size; i++)
 	{
